Fix compile error deadlocked waiting on module
