+++ lib/tx_build/claim_predicates.ex
@@ -2,16 +2,12 @@ defmodule Stellar.TxBuild.ClaimPredicates do
   @moduledoc """
   `ClaimPredicates` struct definition.
   """
-  import Stellar.TxBuild.Validations,
-    only: [
-      validate_predicate_list: 2
-    ]
-
   alias Stellar.TxBuild.ClaimPredicate
   alias StellarBase.XDR.ClaimPredicates
 
   @behaviour Stellar.TxBuild.XDR
 
+  @type validation :: {:ok, any()} | {:error, error()}
   @type predicates :: ClaimPredicate.t() | list(ClaimPredicate.t())
   @type error :: {:error, atom()}
   @type t :: %__MODULE__{value: list(ClaimPredicate.t())}
@@ -33,4 +29,14 @@ defmodule Stellar.TxBuild.ClaimPredicates do
     |> Enum.map(&ClaimPredicate.to_xdr/1)
     |> ClaimPredicates.new()
   end
+
+  @spec validate_predicate_list(predicates :: predicates, predicates :: predicates()) ::
+          validation()
+  def validate_predicate_list(response, []), do: {:ok, response}
+
+  def validate_predicate_list(response, [%ClaimPredicate{} = h | t]),
+    do: validate_predicate_list(response ++ [h], t)
+
+  def validate_predicate_list(_response, _predicates),
+    do: {:error, :invalid_predicate_list_value}
 end
